Fix dir structure to make CI happy
