+++ test/units/modules/network/frr/fixtures/frr_bgp_config
@@ -1,0 +1,15 @@
+!
+router bgp 64496
+ bgp router-id 192.0.2.1
+ bgp log-neighbor-changes
+ neighbor 192.51.100.1 remote-as 64496
+ neighbor 192.51.100.1 timers 120 360
+ neighbor 198.51.100.3 remote-as 64498
+ !
+ address-family ipv4 unicast
+  network 192.0.1.0/24 route-map RMAP_1
+  network 198.51.100.0/24
+  redistribute static metric 100
+  redistribute ospf 223 metric 10
+ exit-address-family
+!
